+++ src/main/java/com/qsoftware/forgemod/objects/block/generator/diesel/DieselGeneratorBlock.java
@@ -1,5 +1,7 @@
 package com.qsoftware.forgemod.objects.block.generator.diesel;
 
+import com.qsoftware.forgemod.objects.block.generator.AbstractGeneratorBlock;
+import com.qsoftware.forgemod.util.MachineTier;
 import net.minecraft.block.AbstractBlock;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.SoundType;
@@ -12,8 +14,6 @@ import net.minecraft.util.SoundEvents;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.IBlockReader;
 import net.minecraft.world.World;
-import com.qsoftware.forgemod.objects.block.generator.AbstractGeneratorBlock;
-import com.qsoftware.forgemod.util.MachineTier;
 
 import javax.annotation.Nullable;
 import java.util.Random;
@@ -34,7 +34,7 @@ public class DieselGeneratorBlock extends AbstractGeneratorBlock {
         // TODO: Unique sound and particles? Copied from BlastFurnaceBlock.
         if (stateIn.get(LIT)) {
             double d0 = (double) pos.getX() + 0.5D;
-            double d1 = (double) pos.getY();
+            double d1 = pos.getY();
             double d2 = (double) pos.getZ() + 0.5D;
             if (rand.nextDouble() < 0.1D) {
                 worldIn.playSound(d0, d1, d2, SoundEvents.BLOCK_BLASTFURNACE_FIRE_CRACKLE, SoundCategory.BLOCKS, 1.0F, 1.0F, false);
