+++ src/main/java/com/qsoftware/forgemod/objects/block/generator/lava/LavaGeneratorBlock.java
@@ -1,5 +1,7 @@
 package com.qsoftware.forgemod.objects.block.generator.lava;
 
+import com.qsoftware.forgemod.objects.block.generator.AbstractGeneratorBlock;
+import com.qsoftware.forgemod.util.MachineTier;
 import net.minecraft.block.AbstractBlock;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
@@ -16,15 +18,13 @@ import net.minecraft.util.math.shapes.VoxelShape;
 import net.minecraft.util.math.shapes.VoxelShapes;
 import net.minecraft.world.IBlockReader;
 import net.minecraft.world.World;
-import com.qsoftware.forgemod.objects.block.generator.AbstractGeneratorBlock;
-import com.qsoftware.forgemod.util.MachineTier;
 
 import javax.annotation.Nullable;
 import java.util.Random;
 
 public class LavaGeneratorBlock extends AbstractGeneratorBlock {
 
-    private static final VoxelShape SHAPE = VoxelShapes.or(Block.makeCuboidShape(1, 0, 1, 15, 3, 15), Block.makeCuboidShape(2, 3, 2, 14, 16, 14) );
+    private static final VoxelShape SHAPE = VoxelShapes.or(Block.makeCuboidShape(1, 0, 1, 15, 3, 15), Block.makeCuboidShape(2, 3, 2, 14, 16, 14));
 
     public LavaGeneratorBlock() {
         super(MachineTier.STANDARD, AbstractBlock.Properties.create(Material.IRON).hardnessAndResistance(6, 20).sound(SoundType.METAL));
@@ -46,7 +46,7 @@ public class LavaGeneratorBlock extends AbstractGeneratorBlock {
         // TODO: Unique sound and particles? Copied from BlastFurnaceBlock.
         if (stateIn.get(LIT)) {
             double d0 = (double) pos.getX() + 0.5D;
-            double d1 = (double) pos.getY();
+            double d1 = pos.getY();
             double d2 = (double) pos.getZ() + 0.5D;
             if (rand.nextDouble() < 0.1D) {
                 worldIn.playSound(d0, d1, d2, SoundEvents.BLOCK_BLASTFURNACE_FIRE_CRACKLE, SoundCategory.BLOCKS, 1.0F, 1.0F, false);
